requirements.txt
tensorflow==2.8.0
tensorflow-object-detection-api
numpy==1.20.3
opencv-python==4.5.5.62
protobuf==3.20.1